Material

Material

Refraction.Material — Type

struct Material{MD<:MaterialData}
    name::String
    materialdata::MD
    wavelength_range::Tuple{Float64,Float64}
end

A material stores:

name::String : The name of the material, which is the path to the material in the database or a custom name when creating your own materials

materialdata::{<:MaterialData} : The material data, i.e the refractive index data and extinction coefficient data

wavelength::Tuple{Float64, Float64} : The wavelength range of the material data

Info

Not every material in the database contains refractive indices and extinction coefficients. Some materials only contain refractive indices, while others only contain extinction coefficients.

The material object acts as a functor to compute the refractive index of the material for a given wavelength

(m::Material)(wavelength::Real; [, default::Real=0.0])

More information about the Functor method can be found here Material(wavelength::Real; default::Real)

Refraction.Material — Method

Material(path::String) -> Union{Material, Vector{Material}

Creates a material from the given path in the refractive index database. This method will throw an ArgumentError if the given path does not correspond to a material in the database.

Arguments

path::String : The path to the material given in the format "<shelf>/<book>/<page>"

Returns

Union{Material, Vector{Material} : The material object(s) created from the given path. The method returns a vector of material objects if the page at the given path contains multiple material datasets.

Refraction.Material — Method

Material(shelf, book, page) -> Union{Material, Vector{Material}}

Provide the shelf, book and page of the material separately. This method then calls Material("$shelf/$book/$page").

Arguments

shelf::String : The shelf of the material

book::String : The book of the material

page::String : The page of the material

Refraction.Material — Method

Material(n::Real [,name::String="unnamed"]) -> Material

Construct a material with a constant index of refraction over the entire wavelength range. By default the material will be given the name "unnamed".

Refraction.Material — Method

Material(shelf, book, page) -> Union{Material, Vector{Material}}

Provide the shelf, book and page of the material separately. This method then calls Material("$shelf/$book/$page").

Arguments

shelf::String : The shelf of the material

book::String : The book of the material

page::String : The page of the material

Material(n::Real, k::Real [, name::String="unnamed"]) -> Material

Construct a material with a constant index of refraction and extinction coefficient over the entire wavelength range. By default the material will be given the name "unnamed".

Refraction.Material — Method

Material(; k::Real [, name::String="unnamed"]) -> Material

Construct a material with a constant index of refraction and extinction coefficient over the entire wavelength range. By default the material will be given the name "unnamed".

Finding Materials

Refraction.findmaterial — Function

findmaterial(name::Regex) -> Set{String}

Search for materials in the database that match the given regular expression. The method returns a set of paths to the materials that startswith the regular expression. The search is case-insensitive.

Arguments

name::Regex : The regular expression that the method will match to the start of the material paths.

findmaterial(name::String) -> Set{String}

Arguments

name::String : The name of the material to search for in the database. The name can be written in the format "<shelf>/<book>/<page>" or using spaces instead of forward slashes.

Null Material

Refraction.NULL_MATERIAL — Constant

const NULL_MATERIAL = Material(NaN, "null")

Material object that defines the null material. The null material is defined as a material with name "null" and a constant refractive index of NaN over the entire wavelength range.

Refraction.isnullmaterial — Function

isnullmaterial(m::Material) -> Bool

Check if the given material is the null material. The method returns true if the material is the null material, otherwise it returns false.

Refractive Index

Refraction.Material — Method

(m::Material)(wavelength::Real; [default::Real=1.0]) -> Real

Functor method to compute the refractive index of the material for a given wavelength. If the wavelength is outside the wavelength range of the material the method will clamp the wavelength to the range. In the case that the material does not contain data on the refractive indices of the material the method will return the default value. The functor method calls the dispersion method.

Arguments

m::Material : The material object for which the refractive index is to be computed.

wavelength::Real : The wavelength at which the refractive index is to be computed given in $\mu m$.

default::Real=1.0 : The default value to return if the material does not contain data on refractive indices

Refraction.dispersion — Function

dispersion(m::Material, wavelength::Real; [default::Real=1.0]) -> Real

Compute the refractive index of the material for a given wavelength. If the wavelength is outside the wavelength range of the material the method will clamp the wavelength to the range. In the case that the material does not contain data on the refractive indices of the material the method will return the default value.

Arguments

m::Material : The material object for which the refractive index is to be computed.

wavelength::Real : The wavelength at which the refractive index is to be computed given in $\mu m$.

default::Real=1.0 : The default value to return if the material does not contain data on refractive indices

Extinction Coefficient

Refraction.extinction — Function

extinction(m::Material, wavelength::Real; [default::Float64=0.0]) -> Real

Compute the extinction coefficient of the material for a given wavelength. If the wavelength is outside the wavelength range of the material the method will clamp the wavelength to the range. In the case that the material does not contain data on the extinction coefficient of the material the method will return the default value.

Arguments

m::Material : The material object for which the extinction coefficient is to be computed.

wavelength::Real : The wavelength at which the extinction coefficient is to be computed given in $\mu m$.

default::Float64=0.0 : The default value to return if the material does not contain data on extinction coefficients

Refraction.transmittance — Function

transmittance(m::Material, wavelength::Real, distance::Real; [default::Real=0.0]) -> Real

Compute the transmittance of the material for a given wavelength and distance

Arguments

m::Material : The material object

wavelength::Real : The wavelength at which the transmittance is to be computed given in $\mu m$.

distance::Real : The distance over which the transmittance is to be computed given in $\mu m$.

default::Real=0.0 : The default value to return for the extinction coefficient if the material does not contain data on extinction coefficients